Patch bug fixes (#74)
* Bug Fixes

Fixed the doorway bug, dropping an empty item slot, and using the highlighted instead of the equipped item. Also gave movable items a weight of 1000 so they can not be picked up.

* Update Update.cpp

* Picking up Walls and Doors

No longer Seg Faults when trying to pick up a wall or door

* Changed dropItem

Item could be dropped on another. This has been fixed

* Fixed Item Dup Bug

items[i] was never being set to NULL. This when the item was picked up again after dropping, another pointer to that item existed.